+++ public/hydrogen/time.h
@@ -22,7 +22,7 @@ uint64_t hydrogen_boot_time(void) __asm__("__hydrogen_boot_time");
  *
  * \return The number of nanoseconds that have elapsed since 1970-01-01T00:00:00Z.
  */
-int64_t hydrogen_get_real_time(void) __asm__("__hydrogen_get_real_time");
+__int128_t hydrogen_get_real_time(void) __asm__("__hydrogen_get_real_time");
 
 /**
  * Set the current real time.
@@ -32,7 +32,7 @@ int64_t hydrogen_get_real_time(void) __asm__("__hydrogen_get_real_time");
  * \param[in] time The number of nanoseconds that have elapsed since 1970-01-01T00:00:00Z.
  * \return 0, if successful; if not, an error code.
  */
-int hydrogen_set_real_time(int64_t time) __asm__("__hydrogen_set_real_time");
+int hydrogen_set_real_time(__int128_t time) __asm__("__hydrogen_set_real_time");
 
 #ifdef __cplusplus
 };
